Fixed Issue #39: Set corresponding parameters when setting a run background.
Added a DefaultRunProperties class with some const members that we can refer to anytime we need to reset certain properties to their defaults.
The rule for switching is that the Width, Height and Name properties will be changed to their default values on any switch (except the name, that will be reset to String.Empty if the new mode is Modern).
When setting a new background, the resolution of each run will be set accordingly. Instead of setting these properties based on the background, we can obviously choose to set them as Double.NaN in order to make WPF scale each control accordingly, but this would not work as we need the Width values a lot in many computations.
Also fixed a computation bug when setting a background image that was different in terms of resolution than the default one.